+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Changelog
  * renamed binding type "puli/asset-mapping" to "puli/asset"
  * removed `$code` arguments of static exception factory methods
  * added `puli asset update` command
+ * added `OVERRIDE` flag to `AssetManager`
  
 * 1.0.0-beta (2015-03-19)